Feat : "Index 페이지 내 Follow Peaople 기능 연동"
Index 페이지 상단에 Follow할 친구를 추천 합니다.
지금은 Session을 적용하지 않고 상위 7명을 추천하도록 구성하였습니다.
Session이 적용된 후는 차차 적용하도록 하겠습니다.

Fixes : #15
